app/requirements.txt
fastapi==0.120.1
Pillow==12.0.0
pydantic==2.12.3
pydantic_settings==2.11.0
pytest==8.4.2
SQLAlchemy==2.0.44
starlette==0.49.0
Mako==1.3.10
MarkupSafe==3.0.3
alembic==1.17.0
uvicorn==0.35.0
psycopg2==2.9.10
httpx==0.28.1
python-multipart==0.0.20